Nest Condition and Conclusion inside RuleResolvers
